requirements.txt
python-telegram-bot[job-queue]==21.6
groq==0.9.0
python-dotenv==1.0.1
